requirements.txt
gpytorch~=1.14
torch~=2.4.1
matplotlib~=3.10.0
qadence~=1.10.2
flake8~=7.1.2
pytest~=8.3.4
lightning~=2.5.0
networkx~=3.4.2
numpy~=1.26.4
scipy~=1.15.1
omegaconf~=2.3.0
hydra-core~=1.3.2
tqdm~=4.67.1
scikit-learn~=1.6.1